Structurize integration tests
