+++ .travis.yml
@@ -45,3 +45,4 @@ after_success:
 
 notifications:
   email: false
+
